Full Restart
Started fresh with new package version, a requierement based authoring workflow and a clearer goal : Remake a very simple version of the tank demo.
So far, movement and aiming work as intended.
Need to upgrade the FaceMouvementDirectionSystem to take into account parenting.
Probably need to update the MoveSystem to split it or make the grounded requirement facultative.